+++ docs/_pages/releases.md
@@ -11,6 +11,8 @@ sidebar:
 ### Fixes
 * Fix assembly loading for NUnit4 and `LifeCycle.InstancePerTestCase` with Microsoft.Testing.Platform - [#362](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/362)
 
+* Fix configuring displayed length of string assertions at `AssertionScope` - [#393](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/393)
+
 ## 9.3.0
 
 ### What's new